Use BCP47 locale for xml:lang when in XML bulk mode
In XML, the xml:lang tag is required to use a locale in BCP47 format[1],
however, msgfmt operates on POSIX locale which are not compliant with
BCP47. This is not an issue when msgfmt is used on one single file, as a
language name can be given by the user in that case.
When operating in bulk mode however, it is not easy to pass a BCP47
name, so msgfmt would need to infer one from its POSIX locale.

Fortunately that is relatively easy to do for 99% of all cases.
This patch makes msgfmt use BCP47 for xml:lang in XML bulk mode, which
makes it follow the IETF recommendation and match other tools such as
itstool.

[1]: https://en.wikipedia.org/wiki/IETF_language_tag